* Moved the logic that loaded a schedule from a config file from the core library out to it's own Configuration library. * Removed contamination of the main library Scheduler that was explicitly referencing config objects instead of interfaces and the instances it was handed
